- Merge example modules - Provide akka http marshalers and unmarshallers
