+++ examples/src/main/scala/com/powerspace/openrtb/examples/rtb/http4s/common/ExampleSerdeModule.scala
@@ -1,4 +1,4 @@
-package com.powerspace.openrtb.examples.rtb.common
+package com.powerspace.openrtb.examples.rtb.http4s.common
 
 import com.powerspace.openrtb.example.ExampleProto
 import com.powerspace.openrtb.json.OpenRtbExtensions.ExtensionRegistry
@@ -7,7 +7,8 @@ import com.powerspace.openrtb.json.{OpenRtbExtensions, SerdeModule}
 object ExampleSerdeModule extends SerdeModule {
   override def nativeRegistry: OpenRtbExtensions.ExtensionRegistry = ExtensionRegistry()
 
-  override def extensionRegistry: OpenRtbExtensions.ExtensionRegistry = ExtensionRegistry()
-    .registerExtension(ExampleProto.impExt)
-    .registerExtension(ExampleProto.bidResponseExt)
+  override def extensionRegistry: OpenRtbExtensions.ExtensionRegistry =
+    ExtensionRegistry()
+      .registerExtension(ExampleProto.impExt)
+      .registerExtension(ExampleProto.bidResponseExt)
 }
