--- openrtb-examples/src/main/protobuf/example.proto
@@ -1,13 +1,0 @@
-syntax = "proto2";
-
-import "openrtb.proto";
-
-package com.powerspace.example;
-
-message BidResponseExt {
-    optional string protocol = 1;
-}
-
-extend com.google.openrtb.BidResponse {
-    optional BidResponseExt bidResponseExt = 1000;
-}
